html/obsolete/requirements-for-implementations/other-elements-attributes-and-apis/document-color-fg.html

<!DOCTYPE html>
<title>document wg color</title>
<link rel="help" href="https://html.spec.whatwg.org/multipage/obsolete.html#dom-document-fgcolor">
<link rel="help" href="https://html.spec.whatwg.org/multipage/obsolete.html#dom-document-linkcolor">
<link rel="help" href="https://html.spec.whatwg.org/multipage/obsolete.html#dom-document-vlinkcolor">
<link rel="help" href="https://html.spec.whatwg.org/multipage/obsolete.html#dom-document-alinkcolor">
<link rel="help" href="https://html.spec.whatwg.org/multipage/obsolete.html#dom-document-bgcolor">
<link rel="help" href="https://html.spec.whatwg.org/multipage/obsolete.html">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body text="red" linkColor="red" vlink="red" alink="red"  bgColor="red">
<script>
test(function() {

  document.linkColor = "red";
  
  assert_equals(document.fgColor, "red");

  assert_equals(document.linkColor, "red");

  assert_equals(document.vlinkColor, "red");

  assert_equals(document.alinkColor, "red");

  assert_equals(document.bgColor, "red");

}, "testing every attribute is red")

test(function() {
  document.body.text = "blue";

  assert_equals(document.fgColor, "blue");
}, "document's fgColor IDL attribute reflects the body's text content attribute");

test(function() {
  document.body.link = "green";

  assert_equals(document.linkColor, "green");
}, "document's linkColor IDL attribute reflects the body's link content attribute");

test(function() {
  document.body.vLink = "blue";

  assert_equals(document.vlinkColor, "blue");
}, "document's vlinkColor IDL attribute reflects the body's vLink content attribute");

test(function() {
  document.body.aLink = "orange";

  assert_equals(document.alinkColor, "orange");
}, "document's alinkColor IDL attribute reflects the body's aLink content attribute");

test(function() {
  document.body.bgColor = "yellow";

  assert_equals(document.bgColor, "yellow");
}, "document's bgColor IDL attribute reflects the body's bgColor content attribute");

test(function(){
   document.body.removeAttribute("text");
   assert_equals(document.fgColor, "");

   document.body.removeAttribute("link");
   assert_equals(document.linkColor, "");

   document.body.removeAttribute("vlink");
   assert_equals(document.vlinkColor, "");

   document.body.removeAttribute("alink");
   assert_equals(document.alinkColor, "");

   document.body.removeAttribute("bgColor");
   assert_equals(document.bgColor, "");
   
}, "Removing the attributes causes them to be reflected as the empty string");

</script>
